Inherit metrics client from base client
